Create PS14P1.py
PS14P1 (from my understanding of the directions Mr. Hull), wanted to make sure that the code ran smoothly with the method that was added to compute the employee's bonus/salary. Which is why I uploaded the code in its original formatting with the addition of this method, to make sure that it runs like it's supposed to.